+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 The structure and content of this file follows [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.20.1] - 2023-10-20
+### Fixed
+- Calling jp.Set on a map or struct with a nil value no longer panics
+  if nil is a valid value for the element being set.
+
 ## [1.20.0] - 2023-10-17
 ### Added
 - Added ojg.Options.FloatFormat to allow float output format to be specified.